fix(glm52): hoist tail seq_len out of the graphed region (review)
- set tail_input/k_tail/v_tail seq_len in host prep every round: a
  replayed graph skips the captured closure, so a stale short bound from
  a previous accept length trips the eager middle's range check (the
  engine-teardown crash seen live)
- fix qwen3 DFlash markov argmax scratch sizing: both sizing sites now
  use markov_step_argmax_partials_len after the 1024-elem tile widening
- parity gate reworked to a serving-shaped schedule: mixed accept
  lengths, bs=2 rounds interleaved between bs=1 rounds, and a
  context_len==8 replay right after short-tail rounds; reverting the
  hoist alone makes the gate fail with the exact live-crash signature
  (source token range [0..16) exceeds seq_len 11)
- split dspark.rs back under the line ceiling: slot state ->
  dspark_slot.rs, test-only helpers -> dspark_test_support.rs
- drop dead set_tail_len; serialize env-mutating tests on ENV_LOCK;
  re-derive the smoke VRAM note (~14 GiB) and drop the stale
  --features glm52 build command
